Добавит валидацию props.
